requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-telegram-bot==20.7
aiohttp==3.9.3
python-dotenv==1.0.1
websockets==12.0